Build analytics_test with chrome driver for now
