+++ resources/testanalytics.xml
@@ -10,7 +10,6 @@
 		</groups>
 		<classes>
 			<class name="com.nci.testcases.Analytics_Test" />
-			<class name="com.nci.testcases.Sandbox" />			
 		</classes>
 	</test>
