+++ src/test/java/com/nci/testcases/Analytics_Test.java
@@ -1,11 +1,14 @@
 package com.nci.testcases;
 
+import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
+import java.net.URLDecoder;
 import java.util.concurrent.TimeUnit;
 import java.util.List;
 
 import org.openqa.selenium.Proxy;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.remote.CapabilityType;
 import org.openqa.selenium.remote.DesiredCapabilities;
@@ -29,6 +32,9 @@ public class Analytics_Test extends BaseClass {
 	// WebDriver driver;
 	AnalyticsLoad analytics;
 
+	// Analytics URL 
+	public static final String TRACKING_SERVER = "nci.122.2o7.net";
+	
 	// HAR data object
 	/** A HAR (HTTP Archive) is a file format that can be used by HTTP monitoring 
 	 * tools to export collected data. 
@@ -45,7 +51,8 @@ public class Analytics_Test extends BaseClass {
 		driver = BrowserManager.startBrowser(browser, pageURL);
 		driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 		analytics = new AnalyticsLoad(driver);
-		setupProxy(driver);
+		// setupProxy(driver);		
+		setupProxy();
 		System.out.println("Analytics setup done");
 	}
 
@@ -54,7 +61,8 @@ public class Analytics_Test extends BaseClass {
 	 * Modified from https://github.com/lightbody/browsermob-proxy#using-with-selenium
 	 * @throws RuntimeException
 	 */
-	public void setupProxy(WebDriver driver) throws RuntimeException {
+	// public void setupProxy(WebDriver driver) throws RuntimeException {	 
+	public void setupProxy() throws RuntimeException {
 	    // start the proxy
 	    BrowserMobProxy proxy = new BrowserMobProxyServer();
 	    proxy.start(0);
@@ -66,6 +74,9 @@ public class Analytics_Test extends BaseClass {
 	    DesiredCapabilities capabilities = new DesiredCapabilities();
 	    capabilities.setCapability(CapabilityType.PROXY, seleniumProxy);
 
+	    // New driver
+	    WebDriver driver = new ChromeDriver(capabilities);
+	    
 	    // enable more detailed HAR capture, if desired (see CaptureType for the complete list)
 	    proxy.enableHarCaptureTypes(CaptureType.REQUEST_CONTENT, CaptureType.RESPONSE_CONTENT);
 
@@ -76,13 +87,25 @@ public class Analytics_Test extends BaseClass {
 	    driver.get(config.getPageURL("HomePage"));
 
 	    // get the HAR data and print to console
-	    // TODO: fix har object - no data is being collected
+	    // TODO: Create logic for different browsers. Either here or create a new method in BrowserManager()
+	    // TODO: Start tracking click events
+	    // TODO: Get the driver to actually quit when done 
+	    // TODO: Smash stuff with hammer
 	    har = proxy.getHar();
 	    List<HarEntry> entries = har.getLog().getEntries();
     	System.out.println("Entry count (debug): " + entries.size());
 	    for (HarEntry entry : entries) {
-	    	System.out.println(entry.getRequest().getUrl());
-	    }	    
+	    	if(entry.getRequest().getUrl().contains(TRACKING_SERVER))
+			{
+	    		String result = entry.getRequest().getUrl();
+	    		try {
+					result = URLDecoder.decode(result, "UTF-8");
+				} catch (Exception e) {
+					result = "bleah";
+				} 
+				System.out.println(result);
+			}
+	    }  
 	    
 		System.out.println("BMP proxy setup done");
 	}
